Rename files for <EmojiButton /> to follow other name patterns
