+++ src/components/emoji-button.js
