+++ src/components/index.js
@@ -2,7 +2,7 @@ export { default as Anchors } from './anchors'
 export { default as Category } from './category'
 export { default as Search } from './search'
 export { default as Skins } from './skins'
-export { default as EmojiButton } from './button'
+export { default as EmojiButton } from './emoji-button'
 export { default as Touchable } from './common/touchable'
 
 export { default as Emoji } from './emoji/emoji'
